TableExtraction/Offline/requirements.txt
camelot-py==0.10.1
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.2
et-xmlfile==1.1.0
ghostscript==0.7
numpy==1.23.0
openpyxl==3.0.10
pandas==1.4.3
pdfminer.six==20220524
pycparser==2.21
PyPDF2==2.3.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
tabulate==0.8.10
tk==0.1.0
typing_extensions==4.2.0